Start fleshing out sparse index caching
